User can only make one upvote per post, PEP8 adjustments

+++ main.py
@@ -6,32 +6,37 @@ import string
 import re
 import os
 import jinja2
+import logging
+from google.appengine.ext import db
 
 SECRET = 'supersecretmission'
 
-from google.appengine.ext import db
 
 template_dir = os.path.join(os.path.dirname(__file__), 'templates')
-jinja_env = jinja2.Environment(loader = jinja2.FileSystemLoader(template_dir),
-                                autoescape=True)
+jinja_env = jinja2.Environment(
+    loader=jinja2.FileSystemLoader(template_dir),
+    autoescape=True)
+
 
 ###########################
 ####### DATABASES #########
 ###########################
-def users_key(group = 'default'):
+def users_key(group='default'):
     return db.Key.from_path('users', group)
 
-def blog_key(name = 'default'):
+
+def blog_key(name='default'):
     return db.Key.from_path('blogs', name)
 
+
 class User(db.Model):
-    name = db.StringProperty(required = True)
-    pw_hash = db.StringProperty(required = True)
+    name = db.StringProperty(required=True)
+    pw_hash = db.StringProperty(required=True)
     email = db.StringProperty()
 
     @classmethod
     def by_id(cls, uid):
-        return User.get_by_id(uid, parent = users_key())
+        return User.get_by_id(uid, parent=users_key())
 
     @classmethod
     def by_name(cls, name):
@@ -39,12 +44,12 @@ class User(db.Model):
         return u
 
     @classmethod
-    def register(cls, name, pw, email = None):
+    def register(cls, name, pw, email=None):
         pw_hash = make_pw_hash(name, pw)
-        return User(parent = users_key(),
-                    name = name,
-                    pw_hash = pw_hash,
-                    email = email)
+        return User(parent=users_key(),
+                    name=name,
+                    pw_hash=pw_hash,
+                    email=email)
 
     @classmethod
     def login(cls, name, pw):
@@ -52,43 +57,51 @@ class User(db.Model):
         if u and valid_pw(name, pw, u.pw_hash):
             return u
 
+
 class Post(db.Model):
     author = db.StringProperty()
-    subject = db.StringProperty(required = True)
-    content = db.TextProperty(required = True)
-    created = db.DateTimeProperty(auto_now_add = True)
-    modified = db.DateTimeProperty(auto_now = True)
+    subject = db.StringProperty(required=True)
+    content = db.TextProperty(required=True)
+    created = db.DateTimeProperty(auto_now_add=True)
+    modified = db.DateTimeProperty(auto_now=True)
     votes = db.IntegerProperty(default=0)
+    voted = db.StringProperty()
 
     def render(self, loggedIn):
         self._render_text = self.content.replace('\n', '<br>')
-        return render_env("post.html", p=self, loggedIn= loggedIn)
+        return render_env("post.html", p=self, loggedIn=loggedIn)
+
 
 class Comment(db.Model):
     author = db.StringProperty()
-    post_id = db.StringProperty(required = True)
-    comment = db.TextProperty(required = True)
-    created = db.DateTimeProperty(auto_now = True)
+    post_id = db.StringProperty(required=True)
+    comment = db.TextProperty(required=True)
+    created = db.DateTimeProperty(auto_now=True)
+
 
 ################################
 #### SECURITY & VALIDATIONS ####
 ################################
-def make_salt(length = 5):
+def make_salt(length=5):
     return ''.join(random.choice(string.letters) for x in xrange(length))
 
-def make_pw_hash(name, pw, salt = None):
+
+def make_pw_hash(name, pw, salt=None):
     if not salt:
         salt = make_salt()
     h = hashlib.sha256(name + pw + salt).hexdigest()
     return '%s,%s' % (salt, h)
 
+
 def make_secure_val(val):
     return '%s|%s' % (val, hmac.new(SECRET, val).hexdigest())
+
 
 def check_secure_val(secure_val):
     val = secure_val.split('|')[0]
     if secure_val == make_secure_val(val):
         return val
+
 
 def valid_pw(name, password, h):
     salt = h.split(',')[0]
@@ -96,20 +109,25 @@ def valid_pw(name, password, h):
 
 USER_RE = re.compile(r"^[a-zA-Z0-9_-]{3,20}$")
 PASS_RE = re.compile(r"^.{3,20}$")
-EMAIL_RE  = re.compile(r'^[\S]+@[\S]+\.[\S]+$')
+EMAIL_RE = re.compile(r'^[\S]+@[\S]+\.[\S]+$')
+
 
 def valid_username(username):
     return username and USER_RE.match(username)
 
+
 def valid_password(password):
     return password and PASS_RE.match(password)
+
 
 def valid_email(email):
     return not email or EMAIL_RE.match(email)
 
+
 def render_env(template, **params):
     t = jinja_env.get_template(template)
     return t.render(params)
+
 
 ##########################
 ##### MASTER HANDLER #####
@@ -147,6 +165,7 @@ class BaseHandler(webapp2.RequestHandler):
         uid = self.read_secure_cookie('user_id')
         self.user = uid and User.by_id(int(uid))
 
+
 ##################################
 ####### POST & COMMENTS ##########
 ##################################
@@ -154,69 +173,83 @@ class MainPageHandler(BaseHandler):
     def get(self):
         posts = db.GqlQuery("SELECT * FROM Post ORDER BY votes DESC LIMIT 10")
         if self.user:
-            self.render('front.html', posts = posts, loggedIn = self.user )
+            self.render('front.html',
+                        posts=posts,
+                        loggedIn=self.user)
         else:
             self.redirect('/signup')
+
 
 class NewPostHandler(BaseHandler):
     def get(self):
         if self.user:
-            self.render("newpost.html", loggedIn = self.user)
+            self.render("newpost.html", loggedIn=self.user)
         else:
             self.redirect("/")
 
     def post(self):
-        author = self.user.name
-        subject = self.request.get('subject')
-        content = self.request.get('content')
-
-        if subject and content:
-            p = Post(parent = blog_key(),
-                     author = author,
-                     subject = subject,
-                     content = content)
-
-            p.put()
-            post_id = str(p.key().id())
-            self.key = post_id
-
-            self.redirect('/')
+        if not self.user:
+            self.redirect("/login")
         else:
-            error = "Please Fill In All Fields"
-            self.render("newpost.html",
-                        author = author,
-                        subject = subject,
-                        content = content,
-                        error = error,
-                        loggedIn = self.user)
+            author = self.user.name
+            subject = self.request.get('subject')
+            content = self.request.get('content')
+
+            if subject and content:
+                p = Post(parent=blog_key(),
+                         author=author,
+                         subject=subject,
+                         content=content)
+
+                p.put()
+                post_id = str(p.key().id())
+                self.key = post_id
+
+                self.redirect('/')
+            else:
+                error = "Please Fill In All Fields"
+                self.render("newpost.html",
+                            author=author,
+                            subject=subject,
+                            content=content,
+                            error=error,
+                            loggedIn=self.user)
+
 
 class NewCommentHandler(BaseHandler):
     def get(self, post_id):
-        key = db.Key.from_path('Post', int(post_id), parent = blog_key())
+        key = db.Key.from_path('Post', int(post_id), parent=blog_key())
         post = db.get(key)
+        if not self.user:
+            self.redirect("/login")
+        else:
+            comments = db.GqlQuery("SELECT * FROM Comment " +
+                                   "WHERE post_id = :1 " +
+                                   "ORDER BY created DESC",
+                                   post_id)
 
-        comments = db.GqlQuery("SELECT * FROM Comment "
-                               + "WHERE post_id = :1 "
-                               + "ORDER BY created DESC",
-                               post_id)
-
-        self.render("newcomment.html", post = post,
-                    comments = comments, loggedIn = self.user)
+            self.render("newcomment.html",
+                        post=post,
+                        comments=comments,
+                        loggedIn=self.user)
 
     def post(self, post_id):
-        key = db.Key.from_path('Post', int(post_id), parent = blog_key())
-        post = db.get(key)
+        if not self.user:
+            self.redirect("/login")
+        else:
+            posts = db.GqlQuery(
+                "SELECT * FROM Post ORDER BY votes DESC LIMIT 10")
+            author = self.user.name
+            comment = self.request.get('comment')
 
-        author = self.user.name
-        comment = self.request.get('comment')
+            if comment:
+                c = Comment(author=author,
+                            post_id=post_id,
+                            comment=comment)
+                c.put()
 
-        if comment:
-            c = Comment(author = author,
-                        post_id = post_id,
-                        comment = comment)
-            c.put()
+            self.render('front.html', posts=posts, loggedIn=self.user)
 
-        self.redirect('/')
 
 class EditPostHandler(BaseHandler):
     def get(self):
@@ -224,32 +257,32 @@ class EditPostHandler(BaseHandler):
             self.redirect("/login")
         else:
             post_id = self.request.get('id')
-            key = db.Key.from_path('Post', int(post_id), parent = blog_key())
+            key = db.Key.from_path('Post', int(post_id), parent=blog_key())
             post = db.get(key)
 
             if self.user.name == post.author:
-                self.render('editpost.html', p = post, loggedIn = self.user)
+                self.render('editpost.html', p=post, loggedIn=self.user)
             else:
                 msg = "You are not authorized to edit this post."
-                self.render('message.html', msg = msg)
+                self.render('message.html', msg=msg)
 
     def post(self):
-        if not self.user:
+        post_id = self.request.get('id')
+        key = db.Key.from_path('Post', int(post_id), parent=blog_key())
+        p = db.get(key)
+
+        if self.user.name != p.author:
             self.redirect("/login")
         else:
-            post_id = self.request.get('id')
             new_content = self.request.get('editpost')
-            key = db.Key.from_path('Post', int(post_id), parent = blog_key())
-            p = db.get(key)
-
             if new_content:
                 p.content = new_content
                 p.put()
                 self.redirect('/')
-                # self.redirect('/%s' % post_id)
             else:
                 error = "Content cannot be empty."
-                self.render("editpost.html", p = p, error = error)
+                self.render("editpost.html", p=p, error=error)
+
 
 class DeletePostHandler(BaseHandler):
     def get(self):
@@ -257,7 +290,7 @@ class DeletePostHandler(BaseHandler):
             self.redirect("/login")
         else:
             post_id = self.request.get('id')
-            key = db.Key.from_path('Post', int(post_id), parent = blog_key())
+            key = db.Key.from_path('Post', int(post_id), parent=blog_key())
             post = db.get(key)
 
             if self.user.name == post.author:
@@ -265,6 +298,7 @@ class DeletePostHandler(BaseHandler):
                 self.redirect('/')
             else:
                 self.redirect('/')
+
 
 class EditCommentHandler(BaseHandler):
     def get(self):
@@ -275,17 +309,18 @@ class EditCommentHandler(BaseHandler):
             key = db.Key.from_path('Comment', int(comment_id))
             comment = db.get(key)
 
-            self.render("editcomment.html", comment = comment, loggedIn = self.user)
+            self.render("editcomment.html",
+                        comment=comment,
+                        loggedIn=self.user)
 
     def post(self):
-        if not self.user:
+        comment_id = self.request.get('id')
+        edit_comment = self.request.get('editcomment')
+        key = db.Key.from_path('Comment', int(comment_id))
+        comment = db.get(key)
+        if self.user.name != comment.author:
             self.redirect("/login")
         else:
-            comment_id = self.request.get('id')
-            edit_comment = self.request.get('editcomment')
-            key = db.Key.from_path('Comment', int(comment_id))
-            comment = db.get(key)
-
             if edit_comment:
                 comment.comment = edit_comment
                 comment.put()
@@ -293,29 +328,46 @@ class EditCommentHandler(BaseHandler):
             else:
                 self.redirect("/editcomment?id="+comment_id)
 
+
 class DeleteCommentHandler(BaseHandler):
     def get(self):
+        comment_id = self.request.get('id')
+        key = db.Key.from_path('Comment', int(comment_id))
+        comment = db.get(key)
         if not self.user:
             self.redirect("/login")
         else:
-            comment_id = self.request.get('id')
-            key = db.Key.from_path('Comment', int(comment_id))
-            comment = db.get(key)
+            if self.user.name != comment.author:
+                self.redirect("/login")
+            else:
+                db.delete(key)
+                self.redirect("/")
 
-            db.delete(key)
-            self.redirect("/")
 
 class UpvoteHandler(BaseHandler):
     def get(self):
         if self.user:
+            logging.info(self.user.name + ',')
             post_id = self.request.get('id')
-            key = db.Key.from_path('Post', int(post_id), parent = blog_key())
+            key = db.Key.from_path('Post', int(post_id), parent=blog_key())
             post = db.get(key)
-            post.votes += 1
-            post.put()
-            self.redirect("/")
+            if post.voted:
+                voters = post.voted.split(',')
+                if self.user.name not in voters:
+                    post.votes += 1
+                    post.voted += (self.user.name + ',')
+                    post.put()
+                    self.redirect("/")
+                else:
+                    self.redirect("/")
+            else:
+                post.votes += 1
+                post.voted = (self.user.name + ',')
+                post.put()
+                self.redirect("/")
         else:
             self.redirect('/login')
+
 
 ###################################
 ########  USER MANAGEMENT  ########
@@ -331,8 +383,8 @@ class SignUpHandler(BaseHandler):
         self.confirm = self.request.get('confirm')
         self.email = self.request.get('email')
 
-        params = dict(username = self.username,
-                      email = self.email)
+        params = dict(username=self.username,
+                      email=self.email)
 
         u = User.by_name(self.username)
         if u:
@@ -363,6 +415,7 @@ class SignUpHandler(BaseHandler):
             self.login(u)
             self.redirect('/')
 
+
 class LoginHandler(BaseHandler):
     def get(self):
         self.render('login-form.html')
@@ -377,7 +430,8 @@ class LoginHandler(BaseHandler):
             self.redirect('/')
         else:
             msg = "Invalid Username/Password"
-            self.render('login-form.html', error = msg)
+            self.render('login-form.html', error=msg)
+
 
 class LogoutHandler(BaseHandler):
     def get(self):
@@ -398,4 +452,4 @@ url_map = [
     ('/logout', LogoutHandler),
 ]
 
-app = webapp2.WSGIApplication(url_map, debug = True)
+app = webapp2.WSGIApplication(url_map, debug=True)
